% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{fb_load}
\alias{fb_load}
\title{Load previously generated mocking functions for all tables in the given db and schema}
\usage{
fb_load(dbname, path = ".", schema_name = "public")
}
\arguments{
\item{dbname}{name of the database}

\item{schema_name}{name of the db schema}
}
\value{
list of named functions corresponding to the tables in the schema
}
\description{
Load previously generated mocking functions for all tables in the given db and schema
}
