% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate file containing mocking functions for all tables in the given db schema}
\usage{
generate(con, schema_name, path)
}
\arguments{
\item{con}{DBI connection}

\item{schema_name}{name of the db schema}

\item{path}{path to the directory where the mocks should be generated}
}
\description{
Generate file containing mocking functions for all tables in the given db schema
}
