% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate mocking functions for all tables in the given db schema}
\usage{
generate(con, schema_name = "public")
}
\arguments{
\item{con}{DBI connection}

\item{schema_name}{name of the db schema}
}
\value{
list of named functions corresponding to the tables in the schema
}
\description{
Generate mocking functions for all tables in the given db schema
}
